Fix issues with slowness and paths
Kubectl's completion takes ages to load when loaded as part of this
plugin. Removing the loading of the completion from the inside of the
plugin and doing it outside with the help of zinit (so it can defer the
loading and speed things up).

Also fixing issues with the plugin paths when used with zinit as it does
not download the plugin with the directory name of kube-aliases.